split out MidForm constprop
